code/requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
click==8.1.3
cryptography==37.0.2
evdev==1.5.0
filelock==3.7.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
homepage2vec==0.0.3rc0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.8.1
hyperframe==5.2.0
idna==2.10
joblib==1.1.0
lxml==4.9.0
nltk==3.6.7
numpy==1.23.0
opencv-python==4.5.4.58
outcome==1.2.0
packaging==21.3
Pillow==8.2.0
pycparser==2.21
pynput==1.7.4
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyphen==0.12.0
python-dateutil==2.8.2
python-xlib==0.31
PyYAML==6.0
regex==2022.6.2
requests==2.25.1
requests-file==1.5.1
rfc3986==1.5.0
scikit-learn==1.1.1
scipy==1.8.1
selenium==4.0.0
sentence-transformers==2.2.2
sentencepiece==0.1.96
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stemming==1.0
textstat==0.7.2
threadpoolctl==3.1.0
tldextract==3.3.0
tokenizers==0.12.1
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
transformers==4.20.1
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==4.2.0
urllib3==1.26.9
wsproto==1.1.0
zxcvbn==4.4.28
